Fix accidentally renamed time module
